Add RSL tree sitter implementation
I worked on the implementation in another repo so I'm just wholesale
importing it here. My commit messages weren't interesting, so nothing
lost!

I'm still figuring things out but from everything I understand, tree
sitter is a very promising tool for our purposes. Right now, rad's lexer
& parser are handwritten and doesn't at all handle things like
incremental parsing, best-effort error parsing (i.e. parsing incomplete
code well as a best effort), and it's not ready to be used elsewhere in
e.g. an LSP server (RLS) or for syntax highlighting.

Rather than reworking it to address these things, which would be quite
challenging to do well (especially the incremental parsing and error
handling), I'm hoping that moving to a tree sitter implementation for
lexing/parsing, we can allow it and its underlying grammar.js to be the
single 'source of truth' for RSL. I can then leverage the tree sitter
impl for other purposes e.g. rad's interpreter, RLS, Textmate bundle
generation, and Pygment generation.

The implementation in this commit is not quite complete. I've left some
things out that require more work in the scanner.c file, namely:

1. Ensuring correct escaping rules, as we've defined them thus far in
   RSL.
2. Multiline strings
3. Switch statements/exprs (I mean to revisit this area of RSL anyway).

I would love to include this stuff in the amter/rad repo and take a
monorepo approach but a couple of forces are against me:

1. Tree sitter itself seems to take the preference of being in its own
   repo, for versioning and publishing.
2. Golang modules and dependencies also seem to work best when you shard
   out into different git repos.

Maybe there's a way to make it work, but I don't want to spend a load of
time on making it work, now.

Hesitant acronym for RSL Tree Sitter is RTS, referring to our tree
sitter implementation. Might see that popping up in future.