Rad sort stmt: Only allow on one line - don't read across lines
The parser had an issue where it would parse the sort stmt in the
following:

  rad url:
    fields f1, f2
    sort
    f1:
      map n -> 2 * n

as if it were 'sort f1' i.e. it'd read the 'f1' from the start of the
next like, due to the prec.right and `\s` being part of the 'extras'
rule.

This was a bit of a headache to figure out how to fix. One approach I
spent a bit of time on was not allowing newlines in 'extras' and then
having to manually specify my newlines in every single rule in the
grammar. This showed *some* promise but greatly complicated a lot of the
grammar, and I kept hitting issues with ambiguity and tree sitter not
being able ot generate. Ultimately I abandoned the approach after
reading some more and seeing other people have faced similar challenges:

 https://github.com/tree-sitter/tree-sitter/discussions/1421
 https://github.com/tree-sitter/tree-sitter/issues/931

These suggest a change to tree sitter that I also sorta felt like I was
needing - the ability to say "within this rule, ignore the 'extras'.

Anyway, from reading these discussions, it appeared like a go-to
workaround was rampant use of token.immediate(), which I've now done in
this commit, and seems to do the trick. I'll probably be re-using this
trick in more places where this sorta issue occurs, that I've not found
yet.