Multiline string fix: Don't read too many characters, preventing interp
A liiittle bit of an edge case:

  a = """
  "{name}"
  """

Prior to this commit, this would only parse as a single string content,
no interpolation. That's because we'd erroneously read the "{ *together*
in one loop, preventing the next loop from properly detecting that an
interpolation was happening, and so we'd continue gobbling up the string
as a content.

The fix is to just re-loop before we consume the '{' after ", so the new
loop can properly detect we're hitting an interpolation and handle
things correctly.