Overhaul exprs a bit
I ran into a relatively simple case of an or expression where a operand
was indexed and this wasn't getting parsed properly.
Specifically:

 a > 1 or b[a] == 1

was being parsed as

 (a > 1 or b)[a] == 1

I explored some ways to fix this, the only one that seemed to work was
what we do in this commit, which is the traditional 'ladder cascade'
approach lower-priority exprs may delegate to higher-prio ones, wrapping
them.

This solves the precedence issue, but results in some unfortunate,
longer CSTs:

  ternary_expr
    delegate: or_expr
      delegate: and_expr
        delegate: compare_expr
          delegate: add_expr
            delegate: mult_expr
              delegate: unary_expr
                delegate: var_path
                  root: identifier `a`

Make human-parsing these syntax trees a little longer and probably has a
slight performance hit, but most important thing here is correctness, so
a trade well-worth making. In future, there might be a way to avoid
this.