Add support for simple multi-assigns
Prior to this commit

  a, b = 1, 2

was not allowed. It *was* allowed in two cases though:

 1. switch exprs
 2. Multi-output functions

Overall, I think this is perhaps a little inconsistent, potentially
surprising. It's also a sort of limitation that a more enterprise-y
language would enforce, the argument being that it's harder to read or
easier to make mistakes. I respect those reasons, and they're why I had
this limitation in the first place, but also, RSL is a scripting
language, and we don't wanna be too dogmatic about these sorts of rules;
it might be prudent & productive to use them once in a while.

This commit edits the grammar to allow these simple multi-assigns.