Remove support for Go-style if 'pre-stmt' syntax
 if ok = try_get(); ok:
    ...

I like this syntax in Go, which is why I added it to RSL as an
experiment. However, I've come to the conclusion that it doesn't make
sense in RSL, primarily because we don't have scoped variables. Any
variable defined at any point is accessible, regardless of where it was
defined (even if it was in a given block). This is a deliberate design
choice for RSL, given its domain is scripting.

So a syntax like this:

  if a; b:
    c

is equivalent to this:

  a
  if b:
    c

Sure it lets you put 'a' and 'b' on the same line, but it really doesn't
offer much value and just complicates the language needlessly. This
commit removes it from the grammar and restores the 'simple' if
statement.