Implement UFCS grammar for function calling
UFCS is basically being able to invoke this:

  foo(a, b, c)

like this instead:

  a.foo(b, c)

This is useful for when you want to chain function calls. For example,
instead of this:

  print(upper(replace(trim(mystring), "l", "o")))

we can now write this:

  mystring.trim().replace("l", "o").upper().print()

soo much better. The first one is actually so bad you'd probably split
the calls across lines as separate assignments.

A common objection to UFCS is it leads to confusion between object
methods vs. UFCS functio calls. This doesn't apply to RSL because we
*don't have* object method calls, so there's no ambiguity, so UFCS is
potentially a good fit. We'll add it to RSL as a bit of an experiment,
though I'm optimistic about it.

This commit updates the grammar to support it.