Adjust lambda position in grammar
We move lambda up *out* of primary_expr so that the grammar doesn't
imply that lambdas are indexable ( _postfix_expr ). By moving them up
into the unary_expr, we can exit early and make it clear that
indexables, if present, should be a part of the lambda itself, rather
than indexing the lambda.