Adjust lambda grammar to be unambiguous for multireturns
The existing grammar was ambigous. For example:

  process(fn() a, b)

Is this giving two args to 'process', a lambda and 'b' ? Or is it
giving a single lambda with itself returns two things, 'a' and 'b'?

It's ambiguous. To resolve this, we simply require that multireturn
lambdas must surround their return with (). So for the two respective
interpretations of the above code, you'd write:

 1. process(fn() a, b)
 2. process(fn() (a, b))