Allow stmts in lambdas; fn blocks + identifiers in rad block maps
We were only allowing expressions, but it'd be nice to allow stmts so
you can have e.g.

  say = fn(text) $!`say {text}`

where that shell cmd doesn't return anything but this is still a useful
function to invoke.

We were already requiring in-line lambda definitiosn in rad block
mapping modifiers, but that's not the intention. We should probably
allow exprs actually so you can use e.g. a ternary, but we just allow
identifiers directly for now.