Change list syntax from e.g. [*int] -> int[], and adjust fn params
I came to the realization that the variadic syntax e.g. 'x: *int' was
not going to work, as there was room for ambiguity in calls like foo(1)
if the typing was e.g. 'fn foo(x: *int|int)' (should the x be 1 or
[1]?).

What other languages do, and I think at least in part for this reason,
is make the variadic asterisk* a modifier of the parameter name rather
than the type i.e. '*x: int'.

That's what we do as well, in this commit.

That means that the [*int] syntax no longer makes sense, so we change
that back to the ol' int[] syntax.

Also, we adjust our function typing param list syntax a bit to make it
more strict at parse time, rather than us needing to check things are
correct at runtime (e.g. for if a variadic parameter is declared after a
named-only divider).