feat: add scientific_number for int args
We want int args to allow scientific notation e.g. 1e10. However, this
is not currently allowed -- only float args can use scientific notation.

It's hard to make this 100% reliable. For example, 1.234e10 is a valid
int, but 1.234e2 is not. Hard to validate this via a grammar.

So we generally allow all scientific notations in int args, and they're
invalid, we'll detect that in our Rad checker.