feat: separate scientific notation from float
The previous commit allows us to specify scientific notation on ints in
the args blocks.

We want the same to be possible for defaults in int arg args to custom
functions.

To accomplish this, we just need to change the float regex such that we
never parse floats into that, and instead always parse them as
scientific notation literals.

Rad Checker will then, at runtime (or when the user calls it) verify
scientific notation is used appropriately e.g. that the scientific
notation actually resolves to an int (and not a float) when used for a
float.